requirements.txt
django==1.11.7
django-crequest==2016.3.16
djangorestframework==3.7.1
django-easy-select2==1.3.4
django-select2-forms
psycopg2